move namespace up, no need to nest
